refactor(handoff): extract PlanToLeadExecutor.js into modular structure
Part of SD-LEO-REFACTOR-PLANTOLEAD-001 (Child 5 of orchestrator SD-LEO-REFACTOR-LARGE-FILES-002)

Refactors 1,427 LOC monolithic PlanToLeadExecutor.js into focused modules:
- gates/prerequisite-check.js - EXEC-TO-PLAN prerequisite validation
- gates/sub-agent-orchestration.js - LEAD_FINAL orchestration
- gates/retrospective-quality.js - Retrospective quality gate
- gates/git-commit-enforcement.js - Git commit enforcement
- gates/traceability-gates.js - Gate 3 and Gate 4
- gates/user-story-existence.js - User story existence gate
- gates/index.js - Gate exports
- state-transitions.js - User story finalization, parent SD completion
- plan-verification.js - Plan verification validation
- remediation.js - Gate remediation messages
- index.js - Main executor class with 17 exports

Testing: Dynamic import verified, ESLint clean
Total: 1,592 LOC in modular structure, original file reduced to 14 LOC re-export
