+++ scripts/modules/handoff/executors/plan-to-lead/gates/index.js
@@ -1,0 +1,13 @@
+/**
+ * Gate Exports Index for PLAN-TO-LEAD
+ * Part of SD-LEO-REFACTOR-PLANTOLEAD-001
+ *
+ * Re-exports all gate creators for PLAN-TO-LEAD handoff
+ */
+
+export { createPrerequisiteCheckGate } from './prerequisite-check.js';
+export { createSubAgentOrchestrationGate } from './sub-agent-orchestration.js';
+export { createRetrospectiveQualityGate } from './retrospective-quality.js';
+export { createGitCommitEnforcementGate } from './git-commit-enforcement.js';
+export { createTraceabilityGate, createWorkflowROIGate, requiresTraceabilityGates } from './traceability-gates.js';
+export { createUserStoryExistenceGate } from './user-story-existence.js';
