+++ scripts/modules/handoff/executors/plan-to-lead/gates/retrospective-quality.js
@@ -1,0 +1,232 @@
+/**
+ * Retrospective Quality Gate for PLAN-TO-LEAD
+ * Part of SD-LEO-REFACTOR-PLANTOLEAD-001
+ *
+ * SD-CAPABILITY-LIFECYCLE-001: Validates retrospective exists AND has quality content
+ */
+
+import { isInfrastructureSDSync, getThresholdProfile } from '../../../../sd-type-checker.js';
+
+/**
+ * Create the RETROSPECTIVE_QUALITY_GATE validator
+ *
+ * @param {Object} supabase - Supabase client
+ * @returns {Object} Gate configuration
+ */
+export function createRetrospectiveQualityGate(supabase) {
+  return {
+    name: 'RETROSPECTIVE_QUALITY_GATE',
+    validator: async (ctx) => {
+      console.log('\n🔒 RETROSPECTIVE QUALITY GATE');
+      console.log('-'.repeat(50));
+
+      // Check for orchestrator children
+      const parentSdId = ctx.sd?.id || ctx.sdId;
+      const { data: children, error: childError } = await supabase
+        .from('strategic_directives_v2')
+        .select('id, title, status')
+        .eq('parent_sd_id', parentSdId);
+
+      if (childError) {
+        console.log(`   ⚠️ Child query error: ${childError.message}`);
+      }
+
+      const isOrchestrator = children && children.length > 0;
+      const allChildrenComplete = isOrchestrator && children.every(c => c.status === 'completed');
+
+      if (isOrchestrator) {
+        console.log(`   📂 Orchestrator SD detected: ${children.length} children`);
+        if (allChildrenComplete) {
+          console.log('   ✅ All children completed - using relaxed threshold (50%)');
+        }
+      }
+
+      // Store orchestrator context for executeSpecific
+      ctx._isOrchestrator = isOrchestrator;
+      ctx._orchestratorChildren = children || [];
+      ctx._isOrchestratorWithAllChildrenComplete = allChildrenComplete;
+
+      // Load retrospective for this SD
+      const sdUuid = ctx.sd?.id || ctx.sdId;
+      const { data: retrospective, error: retroError } = await supabase
+        .from('retrospectives')
+        .select('*')
+        .eq('sd_id', sdUuid)
+        .order('created_at', { ascending: false })
+        .limit(1)
+        .maybeSingle();
+
+      if (retroError && retroError.code !== 'PGRST116') {
+        console.log(`   ⚠️  Retrospective query error: ${retroError.message}`);
+      }
+
+      // Check for auto-pass conditions
+      const autoPassResult = await checkAutoPassConditions(ctx, retrospective, children, allChildrenComplete);
+      if (autoPassResult) return autoPassResult;
+
+      // Standard retrospective quality validation
+      const { validateSDCompletionReadiness, getSDImprovementGuidance } = await import('../../../../sd-quality-validation.js');
+      const retroGateResult = await validateSDCompletionReadiness(ctx.sd, retrospective);
+      ctx._retroGateResult = retroGateResult;
+
+      // Dynamic threshold based on SD type
+      const threshold = await determineThreshold(ctx.sd, allChildrenComplete);
+
+      const passesThreshold = retroGateResult.score >= threshold;
+
+      if (!passesThreshold) {
+        const guidance = getSDImprovementGuidance(retroGateResult);
+
+        // Display actionable improvement suggestions
+        if (retroGateResult.improvements?.length > 0) {
+          console.log('\n📋 ACTIONABLE IMPROVEMENTS TO PASS THIS GATE:');
+          console.log('='.repeat(60));
+          retroGateResult.improvements.forEach((imp, idx) => {
+            console.log(`\n${idx + 1}. [${imp.criterion}] (score: ${imp.score}/10, weight: ${Math.round(imp.weight * 100)}%)`);
+            console.log(`   → ${imp.suggestion}`);
+          });
+          console.log('\n' + '='.repeat(60));
+        }
+
+        return {
+          passed: false,
+          score: retroGateResult.score,
+          max_score: 100,
+          issues: retroGateResult.issues,
+          warnings: retroGateResult.warnings,
+          improvements: retroGateResult.improvements,
+          guidance,
+          remediation: 'Ensure retrospective has non-boilerplate key_learnings and action_items'
+        };
+      }
+
+      // Convert AI issues to advisory warnings (non-blocking)
+      const advisoryWarnings = [
+        ...retroGateResult.warnings,
+        ...(retroGateResult.issues || []).map(i => `[Advisory] ${i}`)
+      ];
+
+      console.log(`✅ Retrospective quality gate passed (${retroGateResult.score}% >= ${threshold}% threshold)`);
+      if (advisoryWarnings.length > 0) {
+        console.log('   Advisory notes (non-blocking):');
+        advisoryWarnings.slice(0, 3).forEach(w => console.log(`   • ${w}`));
+      }
+
+      return {
+        passed: true,
+        score: retroGateResult.score,
+        max_score: 100,
+        issues: [],
+        warnings: advisoryWarnings,
+        details: {
+          ...retroGateResult,
+          is_orchestrator: isOrchestrator,
+          all_children_complete: allChildrenComplete,
+          children: children || [],
+          child_count: children?.length || 0
+        }
+      };
+    },
+    required: true
+  };
+}
+
+/**
+ * Check for auto-pass conditions (orchestrator, database, bugfix)
+ */
+async function checkAutoPassConditions(ctx, retrospective, children, allChildrenComplete) {
+  // ORCHESTRATOR FAST-PATH
+  if (allChildrenComplete && retrospective?.quality_score >= 60 && retrospective?.status === 'PUBLISHED') {
+    console.log(`   ✅ ORCHESTRATOR AUTO-PASS: All ${children.length} children completed + retrospective exists`);
+    console.log(`      Retrospective quality_score: ${retrospective.quality_score}/100`);
+    console.log('      Rationale: Orchestrators coordinate, children produce deliverables');
+
+    return {
+      passed: true,
+      score: retrospective.quality_score,
+      max_score: 100,
+      issues: [],
+      warnings: ['Orchestrator auto-pass: Quality validated via children completion'],
+      details: {
+        orchestrator_auto_pass: true,
+        child_count: children.length,
+        children_completed: children.filter(c => c.status === 'completed').length,
+        children: children,
+        retrospective_id: retrospective.id,
+        retrospective_quality: retrospective.quality_score
+      }
+    };
+  }
+
+  const sdType = (ctx.sd?.sd_type || '').toLowerCase();
+
+  // DATABASE FAST-PATH
+  if (sdType === 'database' && retrospective) {
+    console.log('   🗄️  DATABASE AUTO-PASS: Database SD with retrospective exists');
+    console.log(`      Retrospective quality_score: ${retrospective.quality_score || 0}/100`);
+
+    return {
+      passed: true,
+      score: Math.max(retrospective.quality_score || 60, 60),
+      max_score: 100,
+      issues: [],
+      warnings: ['Database auto-pass: Validated via migration success + DATABASE sub-agent'],
+      details: {
+        database_auto_pass: true,
+        sd_type: sdType,
+        retrospective_id: retrospective.id,
+        retrospective_quality: retrospective.quality_score
+      }
+    };
+  }
+
+  // BUGFIX FAST-PATH
+  if ((sdType === 'bugfix' || sdType === 'bug_fix') && retrospective) {
+    console.log('   🔧 BUGFIX AUTO-PASS: Bugfix SD with retrospective exists');
+    console.log(`      Retrospective quality_score: ${retrospective.quality_score || 0}/100`);
+
+    return {
+      passed: true,
+      score: Math.max(retrospective.quality_score || 50, 50),
+      max_score: 100,
+      issues: [],
+      warnings: ['Bugfix auto-pass: Simple fix validated via git commit evidence'],
+      details: {
+        bugfix_auto_pass: true,
+        sd_type: sdType,
+        retrospective_id: retrospective.id,
+        retrospective_quality: retrospective.quality_score
+      }
+    };
+  }
+
+  return null; // No auto-pass
+}
+
+/**
+ * Determine quality threshold based on SD type
+ */
+async function determineThreshold(sd, allChildrenComplete) {
+  const isInfrastructure = isInfrastructureSDSync(sd);
+  const sdType = sd?.sd_type || sd?.category || 'feature';
+  const isBugfix = sdType === 'bugfix' || sdType === 'bug_fix';
+
+  let threshold;
+  if (allChildrenComplete) {
+    threshold = 50;
+    console.log('   📂 Using orchestrator threshold (50%) - all children complete');
+  } else if (isBugfix) {
+    threshold = 50;
+    console.log(`   🔧 Using bugfix SD threshold (50%) - sd_type='${sdType}'`);
+  } else if (isInfrastructure) {
+    const profile = await getThresholdProfile(sd, { useAI: false });
+    threshold = profile.retrospectiveQuality;
+    console.log(`   🔧 Using infrastructure SD threshold (${threshold}%) - sd_type='${sdType}'`);
+  } else {
+    const profile = await getThresholdProfile(sd, { useAI: false });
+    threshold = profile.retrospectiveQuality;
+    console.log(`   📋 Using standard SD threshold (${threshold}%) - sd_type='${sdType}'`);
+  }
+
+  return threshold;
+}
